pyproject.toml
[project]
name = "ai-based-reporting-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.51.0",
    "fastapi>=0.115.12",
    "matplotlib>=3.10.3",
    "mcp>=1.8.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
